Put `BlockhashQuery` in its own module
